22.11.04
파티클 기반 완료
시간에 따른 이미지 변경 설정필요
시간에 따른 크기변경 설정 필요

+++ DX2D_BabaIsYou Potfolio 22.10.17/Baba_Is_You/DX2D_2206/Framework/Math/Math.cpp
@@ -6,7 +6,7 @@ int GameMath::Random(const int & min, const int & max)
     return rand() % (max - min) + min;
 }
 
-int GameMath::Random(const float & min, const float & max)
+float GameMath::Random(const float & min, const float & max)
 {
     float normal = rand() / (float)RAND_MAX;
