+++ DX2D_BabaIsYou Potfolio 22.10.17/Baba_Is_You/DX2D_2206/Framework/Math/Math.h
@@ -8,7 +8,7 @@ namespace GameMath
     };
 
     int Random(const int & min, const int & max);
-    int Random(const float & min, const float & max);
+    float Random(const float & min, const float & max);
 
     float Clamp(const float & min, const float & max, float value);
